on completed torrent close button fix
this commit resolves Issue #10.

# problem
The close button threw an Exception which stated, that there was no torrent which could be removed by the client. That means that we tried to remove a torrent which doesn't even exist anymore.

# Solution
Since the torrent gets destroyed in the 'done' event, we have to check if that torrent does even exist before trying to remove it.